% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcds.R
\name{addcds}
\alias{addcds}
\title{addcds}
\usage{
addcds(df, influence.object, ID)
}
\arguments{
\item{df}{dataframe}

\item{influence.object}{An object returned from influence.ME's influence()}

\item{ID}{ID name of df}
}
\description{
This function returns a dataframe with cook's distance stats obtained through
influence.ME's influence() function.
}
\examples{
addcds()
}
\keyword{cook's}
\keyword{d,}
\keyword{influence,}
\keyword{influence.ME}
