% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allexcept.R
\name{allexcept}
\alias{allexcept}
\title{allexcept}
\usage{
allexcept(x, except)
}
\arguments{
\item{x}{Character vector of all variables (use ls() or colnames() etc)}

\item{except}{Character vector of variables to exclude}
}
\description{
This function produces the n of a sample or subgroup. Similar to length(), but optimized for tabstat. It can be used outside of tabstat as well.
}
\examples{
allexcept()
}
\keyword{select}
\keyword{variables}
