% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bylabel.R
\name{bylabel.f}
\alias{bylabel.f}
\title{A Cat Function
bylabel.f
This function produces a list of names that corresponds to the combination of by variables in the tabstat function. It is not generally useful outside of that function.}
\usage{
bylabel.f(comb, varlist)
}
\arguments{
\item{comb}{A row of a matrix that represents by variables.}

\item{varlist}{A character vector of the names of by variables.}
}
\description{
A Cat Function
bylabel.f
This function produces a list of names that corresponds to the combination of by variables in the tabstat function. It is not generally useful outside of that function.
}
\examples{
bylabel.f()
}
\keyword{bylabel}
