% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{ipak}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{Character vector of packages.}
}
\description{
This is Steven Worthington's excellent ipak function. I've included it
here just for my personal use and convenience and to avoid loading more
packages than I need.
Original code and Steven's package can be found:
https://gist.github.com/stevenworthington/3178163
"ipak function: install and load multiple R packages.
check to see if packages are installed. Install them if they are not, then
load them into the R session."
}
\examples{
ipak()
}
\keyword{package}
