% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastnumbers.R
\name{lastnumbers}
\alias{lastnumbers}
\title{lastnumbers}
\usage{
lastnumbers(x)
}
\arguments{
\item{x}{A string that ends in numbers}
}
\description{
This function returns the last numbers of a string (that ends in numbers) in numerical form.
}
\examples{
> lastnumbers('2343lkasdALKENC____33,,,srle567')
[1] 567
}
\keyword{numbers,}
\keyword{regexp}
\keyword{strsplit,}
