% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgby.R
\name{lgby}
\alias{lgby}
\title{lgby}
\usage{
lgby(df, x, y, by, facet_var, grid_var, error_bars = TRUE, dodge = 1,
  error_bar_width = 1)
}
\arguments{
\item{df}{A dataframe.}

\item{x}{The variable to appear on the x-axis.}

\item{y}{The variable to appear on the y-axis.}

\item{by}{The grouping variable.}

\item{facet_var}{The facet-grouping variable.}

\item{grid_var}{The grid-grouping variable. When using grid_var, it is
essentially interchangable with facet_var, as together they determine the
order and axis of the grid. grid_var cannot be used independently of
facet_var for obvious reasons.}

\item{error_bars}{Include standard error bars?}

\item{dodge}{Error bar position_dodge}

\item{error_bar_width}{Error bar width}
}
\description{
This function quickly executes the aggregate and ggplot functions to produce
a simple line graph with separate lines based on an indicated group.
}
\examples{
lgby()
}
\keyword{graph}
\keyword{line}
