% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmervis.R
\name{lmervis}
\alias{lmervis}
\title{lmervis}
\usage{
lmervis(df, model, x, xlevels = NULL, byvars = NULL, bylevels = NULL,
  covars = NULL, covars.lv = NULL, cint = FALSE, nsim = 500,
  seed = 1234, outtype = "plot", xcov = NULL, xcov.lv = NULL)
}
\arguments{
\item{df}{Datafram.}

\item{model}{A lmer() model object.}

\item{x}{Variable to appear on the x axis.}

\item{xlevels}{A character vector of levels. "identity" utilizes existing levels.}

\item{byvars}{A list of by variables.}

\item{bylevels}{A list containing the corresponding levels of the by variables. "identity" utilizes existing levels, "sd" returns -1 and +1 SD relative to mean.}

\item{timevary}{Any variable that varies in conjunction with time.}

\item{timevarylevels}{List of levels of timevary variables. Must be in a list and of equal length to x.}

\item{combinedx}{Numerical vector. If the time series (x and timevary) are split in ways that do not overlap, combinedx should be included. It must be a list of equal length to x and timevary.}

\item{combinedxname}{The name of the combinedx variable.}
}
\description{
This function produces line graphs based on lmer() models. It can incorporate multiple slope variables, especially useful for time series analyses.
}
\examples{
n()
}
\keyword{ggplot,}
\keyword{graph}
\keyword{line}
\keyword{lmer,}
