% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_sum.R
\name{multinom_sum}
\alias{multinom_sum}
\title{multinom_sum}
\usage{
multinom_sum(x)
}
\arguments{
\item{x}{multinom model object}
}
\description{
This function produces cleaner output of the multinom function (nnet package)
}
\examples{
multinom_sum
}
\keyword{multinom,}
\keyword{multinomial}
\keyword{regression}
