% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{N}
\alias{N}
\title{N}
\usage{
N(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector or column of a matrix or dataframe.}
}
\description{
This function produces the n of a sample or subgroup. Similar to length(), but optimized for tabstat. It can be used outside of tabstat as well.
}
\examples{
n()
}
\keyword{count}
\keyword{n,}
