% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss.R
\name{nmiss}
\alias{nmiss}
\title{nmiss}
\usage{
nmiss(x)
}
\arguments{
\item{x}{Individual observation (produced by idmiss)}
}
\description{
This simple function outputs the number of missing scale items for each observation summarized by idmiss.
}
\examples{
n()
}
\keyword{count}
\keyword{n,}
