% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcorr.R
\name{pwcorr}
\alias{pwcorr}
\title{pwcorr}
\usage{
pwcorr(data, x)
}
\arguments{
\item{data}{The dataset.}

\item{x}{A character vector of variables to be included in the correlation matrix.}
}
\description{
This function reorganizes the rcorr function in the Hmisc package in a way that more closely resembles the pwcorr Stata function. It produces a single table with r, n, and p values, omitting the upper triangle.
}
\examples{
pwcorr()
}
\keyword{correlation}
