% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resort.R
\name{resort}
\alias{resort}
\title{resort}
\usage{
resort(output, nby)
}
\arguments{
\item{output}{tabstat output, must be specific (not the whole list).}

\item{nby}{The number of by-variables in the tabstat output.}
}
\description{
This function reorganizes tabstat output to be more readale. Temporary measure.
}
\examples{
resort
}
\keyword{statistics,}
\keyword{summary}
\keyword{tabstat}
