% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundan.R
\name{roundan}
\alias{roundan}
\title{roundan}
\usage{
roundan(x, roundto)
}
\arguments{
\item{x}{data frame or a list of data frames}

\item{roundto}{digits to round to}
}
\description{
This function will round all numeric columns of a data frame or list of data frames.
}
\examples{
roundan()
}
\keyword{round}
