% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{SEM}
\alias{SEM}
\title{SEM}
\usage{
SEM(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}
}
\description{
This function returns the standard error of the mean.
}
\examples{
Sem()
}
\keyword{error}
\keyword{sem,}
\keyword{standard}
