% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slist.R
\name{slist}
\alias{slist}
\title{slist}
\usage{
slist(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
This function splits a singular character vector separated by spaces into distinct elements. i.e. slist('a b c d e') will output c('a', 'b', 'c', 'd', 'e')
}
\examples{
slist()
}
\keyword{character}
\keyword{split}
\keyword{vector,}
