% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdcent.R
\name{stdcent}
\alias{stdcent}
\title{stdcent}
\usage{
stdcent(df, vars, outtype = "both", merged = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{vars}{character vector of variables to be transformed}

\item{outtype}{standardized (std), centered (cent), or both (both)}

\item{merged}{specifies whether output is merged dataset or just selected transformations (in a dataframe)}
}
\description{
This function produces centered and/or standardized variables. Output is either full dataset with standardized/centered vars or just the selected vars in a dataframe.
}
\examples{
stdcent()
}
\keyword{centered,}
\keyword{standardized,}
\keyword{transformation}
