% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstat.R
\name{tabstat}
\alias{tabstat}
\title{tabstat}
\usage{
tabstat(mydata, myvars, mystats, byvars, roundto = NULL)
}
\arguments{
\item{mydata}{A dataframe.}

\item{myvars}{A character vector of variables to be summarized.}

\item{mystats}{A list of statistics to be included in the summary. NOTE: This must be included in the command as list() and cannot be a pre-existing object.}

\item{byvars}{A character vector of by variables (i.e. variables representing groups or subgroups to be included in the summary).}

\item{roundto}{digits to round to, default does not specify}
}
\description{
This function produces descriptive statistics with a flexibility similar to that of the Stata tabstat function. It can produce statstics of any number of groups and subgroups. It can return any percentile value when using "ptileXX.XX" (quotes are required).
}
\examples{
tabstat
}
\keyword{statistics,}
\keyword{summary}
\keyword{tabstat}
