% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcontain.R
\name{varcontain}
\alias{varcontain}
\title{varcontain}
\usage{
varcontain(df, x, ignorecase = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{x}{A string that the returned variables in the dataframe will contain}
}
\description{
This function returns any variable names in a given data set that contain
the a provided string. The
}
\examples{
tdf <- data.frame(x=4,t=6,e=2,b=4,yep=9,Nope=10,cort1=3,cort4=5,Cort_55=9)
varcontain(tdf,'cort')
[1] "cort1"   "cort4"   "Cort_55"
}
\keyword{names}
\keyword{variable}
