% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcentlong.R
\name{zcentlong}
\alias{zcentlong}
\title{zcentlong}
\usage{
zcentlong(df, vars, id, outtype = "both", merged = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{vars}{character vector of variables to be transformed}

\item{id}{Name of grouping variable.}

\item{outtype}{standardized (std), centered (cent), or both (both)}

\item{merged}{specifies whether output is merged dataset or just selected
transformations (in a dataframe): MUST BE TRUE FOR zcentlong.}
}
\description{
This function produces centered and/or standardized variables for long format
dataframes. Merged must=TRUE for now.
}
\examples{
zcent()
}
\keyword{centered,}
\keyword{standardized,}
\keyword{transformation}
