% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss.R
\name{nmiss}
\alias{nmiss}
\title{nmiss}
\usage{
nmiss(df, vars)
}
\arguments{
\item{df}{Dataframe that has the scale items}

\item{vars}{Character list (c()) indicating names of individual scale items.}
}
\description{
This function outputs the number of missing items for a specified list of
variables.
}
\examples{
nmiss()
}
\keyword{NA}
\keyword{count,}
\keyword{missing,}
\keyword{n,}
